Chore: Drain tests

+++ cseme_lib/src/traits/formwork.rs
@@ -1,3 +1,3 @@
 pub(crate) trait AreaOfFormwork {
-    fn get_area_of_formwork(&self) -> f32;
+    fn get_area_of_formwork(&self) -> f64;
 }
